GridController.cs (позицию тайла берем у контроллера по индексу тайла); InputController.cs (простой интерфейс для убодной установки ввода); PlaceController.cs (тайлы меняют позицию, создаются новый на месте уничтоженных)

+++ TestGame/Domain/BaseObject.cs
@@ -61,7 +61,7 @@ namespace TestGame.Domain
 			_color.Toggle();
 		}
 
-		public virtual void SetPosition(int x, int y)
+		public virtual void SetPosition(float x, float y)
 		{
 			_position.X = x;
 			_position.Y = y;
